requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
pymongo
passlib[bcrypt]
PyJWT
python-dotenv
requests